# openapi/openapi.json
{
  "components": {
    "responses": {
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authenticated, but the credential may not act on this resource. Note: cross-TENANT access returns 404, not 403, so the API is not an existence oracle."
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Missing, malformed or expired credential."
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "description": "Stable refusal kind (deny_* / invalid_credentials / …).",
            "type": "string"
          },
          "reason": {
            "description": "Single-line, secret-free human explanation.",
            "type": "string"
          }
        },
        "required": [
          "error",
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A SaaS session token (login) or a tenant API key. The tenant context is the credential's own organization unless the caller is a platform administrator.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Multi-tenant SaaS contract over the TASK 7A control plane. Trading truth (orders, fills, positions, risk, ledger, HA) remains owned by the TASK 1–6 engines and is exposed read-only to tenants. Billing/usage reads are served through the deterministic exports (`/api/saas/exports?kind=subscription|usage|...`); there are no separate plan-catalogue or usage endpoints.",
    "title": "Sniper Suite SaaS Control Plane",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tenant/backtests/{id}": {
      "get": {
        "description": "One backtest run including `result_json`. Synthetic-price runs are labelled `synthetic` in the stored result; the API never presents them as live performance.",
        "operationId": "tenant.getBacktest",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "The backtest run."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "tags": [
          "tenant"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "the control-plane deployment itself",
      "url": "/"
    }
  ],
...
}